light++, shadow++, cone formula

+++ src/vector_color.c
@@ -6,7 +6,7 @@
 /*   By: rtacos <rtacos@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/16 18:50:57 by rtacos            #+#    #+#             */
-/*   Updated: 2020/09/25 20:02:55 by rtacos           ###   ########.fr       */
+/*   Updated: 2020/09/27 20:45:27 by rtacos           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,26 +37,26 @@ t_color	mult_colors(t_color v1, t_color v2, int minus)
 
 t_color		brightness_change(t_color color, float mult)
 {
-	float	tmp;
+	int	tmp;
 
-	if (((tmp = (float)color.r * mult)) > 255.0)
+	if (((tmp = (int)((float)color.r * mult))) > 255)
 		color.r = 255;
 	else if (tmp <= 0.0)
 		color.r = 0;
 	else
-		color.r = (int)tmp;
-	if (((tmp = (float)color.g * mult)) > 255.0)
+		color.r = tmp;
+	if (((tmp = (int)((float)color.g * mult))) > 255)
 		color.g = 255;
 	else if (tmp <= 0.0)
 		color.g = 0;
 	else
-		color.g = (int)tmp;
-	if (((tmp = (float)color.b * mult)) > 255.0)
+		color.g = tmp;
+	if (((tmp = (int)((float)color.b * mult))) > 255)
 		color.b = 255;
 	else if (tmp <= 0.0)
 		color.b = 0;
 	else
-		color.b = (int)tmp;
+		color.b = tmp;
 	return (color);
 }
